feat(workflow_runs): Add Schema and initial structure
