+++ tap_github/gh_client.py
@@ -151,19 +151,6 @@ def generate_pr_commit_schema(commit_schema):
     return pr_commit_schema
 
 
-def load_schemas():
-    schemas = {}
-
-    for filename in os.listdir(get_abs_path('schemas')):
-        path = get_abs_path('schemas') + '/' + filename
-        file_raw = filename.replace('.json', '')
-        with open(path) as file:
-            schemas[file_raw] = json.load(file)
-
-    schemas['pr_commits'] = generate_pr_commit_schema(schemas['commits'])
-    return schemas
-
-
 def write_metadata(mdata, values, breadcrumb):
     mdata.append(
         {
